* Move StopWatch to mmpose/utils/timer.py * Move effect functions to mmpose/core/visualization/effects.py * Add pose tracking to smooth the keypoint * Add CPU and memory info collection (rely on psutil) * Optimize visualization
