+++ mmpose/core/visualization/__init__.py
@@ -1,3 +1,7 @@
+from .effects import apply_bugeye_effect, apply_sunglasses_effect
 from .image import imshow_bboxes, imshow_keypoints, imshow_keypoints_3d
 
-__all__ = ['imshow_keypoints', 'imshow_keypoints_3d', 'imshow_bboxes']
+__all__ = [
+    'imshow_keypoints', 'imshow_keypoints_3d', 'imshow_bboxes',
+    'apply_bugeye_effect', 'apply_sunglasses_effect'
+]
