+++ mmpose/utils/__init__.py
@@ -1,4 +1,5 @@
 from .collect_env import collect_env
 from .logger import get_root_logger
+from .timer import StopWatch
 
-__all__ = ['get_root_logger', 'collect_env']
+__all__ = ['get_root_logger', 'collect_env', 'StopWatch']
